requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.0
duckdns-updater==1.0.0
idna==3.4
Pillow==10.0.1
pystray==0.19.5
requests==2.31.0
six==1.16.0
urllib3==2.0.6